requirements.txt
pandas>=1.3.0
Office365-REST-Python-Client>=2.3.12
tkcalendar>=1.6.1
flask>=2.0.0
python-dotenv>=0.19.0
ttkbootstrap>=1.10.0
pillow>=9.0.0
office365-rest-python-client>=2.3.12
tkcalendar>=1.6.1
